[ME-04] refactor: extracted fake classes to another files
